# docs/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "法律智能问答系统 API",
        "title": "LexVeritas API",
        "contact": {
            "name": "LexVeritas nsxzhou",
            "email": "1790146932@qq.com"
        },
        "version": "1.0.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/auth/login": {
            "post": {
                "description": "使用邮箱和密码登录系统",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "认证"
                ],
                "summary": "用户登录",
                "parameters": [
                    {
                        "description": "登录请求",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_lexveritas_lex-veritas-backend_internal_dto.LoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "登录成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_lexveritas_lex-veritas-backend_internal_pkg_response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/github_com_lexveritas_lex-veritas-backend_internal_dto.LoginResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/github_com_lexveritas_lex-veritas-backend_internal_pkg_response.Response"
                        }
                    },
                    "401": {
                        "description": "邮箱或密码错误",
                        "schema": {
                            "$ref": "#/definitions/github_com_lexveritas_lex-veritas-backend_internal_pkg_response.Response"
                        }
                    },
                    "403": {
                        "description": "账户被禁用或锁定",
                        "schema": {
                            "$ref": "#/definitions/github_com_lexveritas_lex-veritas-backend_internal_pkg_response.Response"
                        }
                    }
                }
            }
        },
        "/auth/login/phone": {
            "post": {
                "description": "使用手机号和验证码登录系统",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "认证"
                ],
                "summary": "手机验证码登录",
                "parameters": [
                    {
                        "description": "手机登录请求",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_lexveritas_lex-veritas-backend_internal_dto.PhoneLoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "登录成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_lexveritas_lex-veritas-backend_internal_pkg_response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/github_com_lexveritas_lex-veritas-backend_internal_dto.LoginResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/github_com_lexveritas_lex-veritas-backend_internal_pkg_response.Response"
                        }
                    },
                    "401": {
                        "description": "验证码错误",
                        "schema": {
                            "$ref": "#/definitions/github_com_lexveritas_lex-veritas-backend_internal_pkg_response.Response"
                        }
                    }
                }
            }
        },
        "/auth/logout": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "登出当前用户，使访问令牌失效",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "认证"
                ],
                "summary": "用户登出",
                "responses": {
                    "200": {
                        "description": "登出成功",
                        "schema": {
                            "$ref": "#/definitions/github_com_lexveritas_lex-veritas-backend_internal_pkg_response.Response"
                        }
                    }
                }
            }
        },
        "/auth/me": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "获取当前登录用户的详细信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "认证"
                ],
                "summary": "获取当前用户",
                "responses": {
                    "200": {
                        "description": "获取成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_lexveritas_lex-veritas-backend_internal_pkg_response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/github_com_lexveritas_lex-veritas-backend_internal_dto.UserResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "未授权",
                        "schema": {
                            "$ref": "#/definitions/github_com_lexveritas_lex-veritas-backend_internal_pkg_response.Response"
                        }
                    },
                    "404": {
                        "description": "用户不存在",
                        "schema": {
                            "$ref": "#/definitions/github_com_lexveritas_lex-veritas-backend_internal_pkg_response.Response"
                        }
                    }
                }
            }
        },
        "/auth/oauth/{provider}": {
            "get": {
                "description": "重定向到第三方OAuth授权页面",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "认证"
                ],
                "summary": "OAuth 第三方登录",
                "parameters": [
                    {
                        "enum": [
                            "google",
                            "github",
                            "wechat"
                        ],
                        "type": "string",
                        "description": "OAuth提供商",
                        "name": "provider",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "302": {
                        "description": "重定向到OAuth授权页面"
                    },
                    "501": {
                        "description": "OAuth登录尚未实现",
                        "schema": {
                            "$ref": "#/definitions/github_com_lexveritas_lex-veritas-backend_internal_pkg_response.Response"
                        }
                    }
                }
            }
        },
        "/auth/oauth/{provider}/callback": {
            "get": {
                "description": "处理OAuth授权回调,完成登录",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "认证"
                ],
                "summary": "OAuth 回调",
                "parameters": [
                    {
                        "enum": [
                            "google",
                            "github",
                            "wechat"
                        ],
                        "type": "string",
                        "description": "OAuth提供商",
                        "name": "provider",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "授权码",
                        "name": "code",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "状态码",
                        "name": "state",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "登录成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_lexveritas_lex-veritas-backend_internal_pkg_response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/github_com_lexveritas_lex-veritas-backend_internal_dto.LoginResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "参数错误",
                        "schema": {
                            "$ref": "#/definitions/github_com_lexveritas_lex-veritas-backend_internal_pkg_response.Response"
                        }
                    },
                    "501": {
                        "description": "OAuth回调尚未实现",
                        "schema": {
                            "$ref": "#/definitions/github_com_lexveritas_lex-veritas-backend_internal_pkg_response.Response"
                        }
                    }
                }
            }
        },
        "/auth/refresh": {
            "post": {
                "description": "使用刷新令牌获取新的访问令牌",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "认证"
                ],
                "summary": "刷新令牌",
                "parameters": [
                    {
                        "description": "刷新令牌请求",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_lexveritas_lex-veritas-backend_internal_dto.RefreshRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "刷新成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_lexveritas_lex-veritas-backend_internal_pkg_response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/github_com_lexveritas_lex-veritas-backend_internal_dto.TokenPair"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/github_com_lexveritas_lex-veritas-backend_internal_pkg_response.Response"
                        }
                    },
                    "401": {
                        "description": "刷新令牌无效或已过期",
                        "schema": {
                            "$ref": "#/definitions/github_com_lexveritas_lex-veritas-backend_internal_pkg_response.Response"
                        }
                    }
                }
            }
        },
        "/auth/register": {
            "post": {
                "description": "使用邮箱注册新用户，需先获取验证码",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "认证"
                ],
                "summary": "用户注册",
                "parameters": [
                    {
                        "description": "注册请求",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_lexveritas_lex-veritas-backend_internal_dto.RegisterRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "注册成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_lexveritas_lex-veritas-backend_internal_pkg_response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/github_com_lexveritas_lex-veritas-backend_internal_dto.UserResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "请求参数错误或验证码无效",
                        "schema": {
                            "$ref": "#/definitions/github_com_lexveritas_lex-veritas-backend_internal_pkg_response.Response"
                        }
                    },
                    "429": {
                        "description": "验证码尝试次数过多",
                        "schema": {
                            "$ref": "#/definitions/github_com_lexveritas_lex-veritas-backend_internal_pkg_response.Response"
                        }
                    }
                }
            }
        },
        "/auth/send-code": {
            "post": {
                "description": "发送邮箱验证码，用于注册或重置密码",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "认证"
                ],
                "summary": "发送验证码",
                "parameters": [
                    {
                        "description": "发送验证码请求",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_lexveritas_lex-veritas-backend_internal_dto.SendCodeRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "验证码发送成功",
                        "schema": {
                            "$ref": "#/definitions/github_com_lexveritas_lex-veritas-backend_internal_pkg_response.Response"
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/github_com_lexveritas_lex-veritas-backend_internal_pkg_response.Response"
                        }
                    },
                    "429": {
                        "description": "请求过于频繁",
                        "schema": {
                            "$ref": "#/definitions/github_com_lexveritas_lex-veritas-backend_internal_pkg_response.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_lexveritas_lex-veritas-backend_internal_dto.LoginRequest": {
            "type": "object",
            "required": [
                "email",
                "password"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "github_com_lexveritas_lex-veritas-backend_internal_dto.LoginResponse": {
            "type": "object",
            "properties": {
                "token": {
                    "$ref": "#/definitions/github_com_lexveritas_lex-veritas-backend_internal_dto.TokenPair"
                },
                "user": {}
            }
        },
        "github_com_lexveritas_lex-veritas-backend_internal_dto.PhoneLoginRequest": {
            "type": "object",
            "required": [
                "code",
                "phone"
            ],
            "properties": {
                "code": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                }
            }
        },
        "github_com_lexveritas_lex-veritas-backend_internal_dto.RefreshRequest": {
            "type": "object",
            "required": [
                "refreshToken"
            ],
            "properties": {
                "refreshToken": {
                    "type": "string"
                }
            }
        },
        "github_com_lexveritas_lex-veritas-backend_internal_dto.RegisterRequest": {
            "type": "object",
            "required": [
                "code",
                "email",
                "name",
                "password"
            ],
            "properties": {
                "code": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "name": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 2
                },
                "password": {
                    "type": "string",
                    "minLength": 8
                },
                "phone": {
                    "type": "string"
                }
            }
        },
        "github_com_lexveritas_lex-veritas-backend_internal_dto.SendCodeRequest": {
            "type": "object",
            "required": [
                "email",
                "purpose"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "purpose": {
                    "type": "string",
                    "enum": [
                        "register",
                        "reset_password"
                    ]
                }
            }
        },
        "github_com_lexveritas_lex-veritas-backend_internal_dto.TokenPair": {
            "type": "object",
            "properties": {
                "accessToken": {
                    "type": "string"
                },
                "expiresIn": {
                    "description": "秒",
                    "type": "integer"
                },
                "refreshToken": {
                    "type": "string"
                },
                "tokenType": {
                    "description": "Bearer",
                    "type": "string"
                }
            }
        },
...
        "github_com_lexveritas_lex-veritas-backend_internal_dto.UserResponse": {
            "type": "object",
            "properties": {
                "avatar": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "lastLoginAt": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "role": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "tokenQuota": {
                    "type": "integer"
                },
                "tokenUsed": {
                    "type": "integer"
                }
            }
        },
        "github_com_lexveritas_lex-veritas-backend_internal_pkg_response.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "message": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "Bearer": {
            "description": "输入格式: Bearer {token}",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}